+++ cnc_ctrl_v1/Encoder.cpp
@@ -1,5 +1,5 @@
 
-#include "maslow.h"
+#include "Maslow.h"
 
 // Yes, all the code is in the header file, to provide the user
 // configure options with #define (before they include it), and
