+++ cnc_ctrl_v1/Kinematics.cpp
@@ -20,7 +20,7 @@ The Kinematics module relates the lengths of the chains to the position of the c
 in X-Y space.
 */
 
-#include "maslow.h"
+#include "Maslow.h"
 
 
 Kinematics::Kinematics(){
