+++ cnc_ctrl_v1/System.cpp
@@ -17,7 +17,7 @@ Copyright 2014-2017 Bar Smith*/
 
 // This file contains system level functions and states
 
-#include "maslow.h"
+#include "Maslow.h"
 
 bool machineReady(){
   bool ret = false;
