+++ cnc_ctrl_v1/TimerOne.cpp
@@ -36,7 +36,7 @@
 #ifndef TIMERONE_cpp
 #define TIMERONE_cpp
 
-#include "maslow.h"
+#include "Maslow.h"
 
 TimerOne Timer1;              // preinstatiate
