+++ cnc_ctrl_v1/cnc_ctrl_v1.ino
@@ -12,7 +12,7 @@
     
     Copyright 2014-2017 Bar Smith*/
     
-#include "maslow.h"
+#include "Maslow.h"
 #include "CNC_Functions.h"
 
 // Define system global state structure
